Always allow to clone global dashboard queries
