+++ client/src/lib/Queries/QueryDesigner.svelte
@@ -28,7 +28,8 @@
     SEARCHTYPES,
     generateQueryString,
     type Search,
-    type Column
+    type Column,
+    proposeName
   } from "$lib/Queries/query";
   import ErrorMessage from "$lib/Errors/ErrorMessage.svelte";
   import { getErrorDetails, type ErrorDetails } from "$lib/Errors/error";
@@ -255,41 +256,6 @@
       dashboard: result.dashboard,
       role: result.role
     };
-  };
-
-  /**
-   * Takes the list of existing queries, looks for already given clones and returns a proper name.
-   * Expamples:
-   *
-   * For non existing clones
-   *
-   * Monat -> Monat (1)
-   * Monat (1) -> Monat (1) (1)
-   *
-   * Say there is already a clone
-   *
-   * Monat and Monat (1) -> Monat (2)
-   * Monat (1) and Monat (1) (1) -> Monat (1) (2)
-   * Monat (1) (2) and Monat (1) (1) -> Monat (1) (3)
-   *
-   * And so on.
-   *
-   * @param result list of queries
-   * @param name name of the query
-   */
-  const proposeName = (result: any, name: string) => {
-    const clones = result
-      .filter((r: any) => {
-        const re = new RegExp(name.replaceAll("(", "\\(").replaceAll(")", "\\)") + " \\(\\d+\\)");
-        return re.test(r.name);
-      })
-      .map((r: any) => {
-        return r.name;
-      })
-      .sort((a: string, b: string) => a.localeCompare(b, "en", { numeric: true }));
-    if (clones.length === 0) return `${name} (1)`;
-    const highestIndex = parseInt(clones[clones.length - 1].split(name + " (")[1]);
-    return `${name} (${highestIndex + 1})`;
   };
 
   onMount(async () => {
