+++ client/src/lib/Queries/query.ts
@@ -198,9 +198,45 @@ const saveStoredQuery = (query: Query, method: string) => {
   return request(`/api/queries${path}`, method, formData);
 };
 
+/**
+ * Takes the list of existing queries, looks for already given clones and returns a proper name.
+ * Expamples:
+ *
+ * For non existing clones
+ *
+ * Monat -> Monat (1)
+ * Monat (1) -> Monat (1) (1)
+ *
+ * Say there is already a clone
+ *
+ * Monat and Monat (1) -> Monat (2)
+ * Monat (1) and Monat (1) (1) -> Monat (1) (2)
+ * Monat (1) (2) and Monat (1) (1) -> Monat (1) (3)
+ *
+ * And so on.
+ *
+ * @param queries list of queries
+ * @param name name of the query
+ */
+const proposeName = (queries: Query[], name: string) => {
+  const clones = queries
+    .filter((r: any) => {
+      const re = new RegExp(name.replaceAll("(", "\\(").replaceAll(")", "\\)") + " \\(\\d+\\)");
+      return re.test(r.name);
+    })
+    .map((r: any) => {
+      return r.name;
+    })
+    .sort((a: string, b: string) => a.localeCompare(b, "en", { numeric: true }));
+  if (clones.length === 0) return `${name} (1)`;
+  const highestIndex = parseInt(clones[clones.length - 1].split(name + " (")[1]);
+  return `${name} (${highestIndex + 1})`;
+};
+
 export {
   generateQueryString,
   createStoredQuery,
+  proposeName,
   updateStoredQuery,
   COLUMNS,
   ORDERDIRECTIONS,
